refactor(viewer): decompose ImageRendererBase into partials + TrackSlider; fold mini viewer into the full viewer
Phase F: split the ~3,150-line ImageRendererBase<TSurface> into 11 partial-class files by concern (Layout/Toolbar/FileList/Overlays/Histogram/InfoPanel/StatusBar/Transport/Input + the core) and extract one TrackSlider helper so the WB sliders, the 6 wavelet sliders, and the SER transport scrub share a single render + drag-math primitive (geometry byte-identical to the old triplicated sites). Pure file organisation + one behaviour-preserving dedup.

Consolidation (#29): delete VkMiniViewerWidget / IMiniViewerWidget / MiniViewerState; the Live Session preview, polar-align, and guide-cam now host the shared full viewer configured chromeless. New LiveFramePreviewSource : IPreviewSource keeps the subsampled median/MAD stretch-stats trick + the freeze lock (PerChannelBackground sized to the channel count so the renderer's post-stretch background never crashes); ViewerState gains HideChrome + FreezeStretchStats; ImageRendererBase gains OverrideWcs + SetSurfaceSize.
